+++ src/core/ggsw/ggsw_key.c
@@ -5,7 +5,6 @@
 
 #include "glwe_key.h"
 #include "rng.h"
-#include "spqlios_alias.h"
 #include "univariate_polynomial.h"
 #include "utils.h"
